multi-agent-financial-research/pyproject.toml
[project]
name = "multi-agent-financial-research"
version = "0.1.0"
description = "Multi-Agent Financial News System for Mag 7 Stocks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-tavily>=0.2.0",
    "streamlit>=1.28.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.hatch.build.targets.wheel]
packages = ["."]
